requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
astunparse==1.6.3
bleach==6.1.0
blinker==1.8.2
boto3==1.35.7
botocore==1.35.7
bytes==0.0.1.dev4
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
executing==2.0.1
fastapi==0.112.2
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
idna==3.7
intel-openmp==2021.4.0
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kaggle==1.6.17
keras==3.4.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.4.0
mpmath==1.3.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
parso==0.8.4
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==6.0.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
redis==5.0.7
requests==2.32.3
rich==13.7.1
s3transfer==0.10.2
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
setuptools==70.3.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
spotipy==2.24.0
SQLAlchemy==2.0.31
stack-data==0.6.3
starlette==0.38.2
sympy==1.13.0
tbb==2021.13.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
text-unidecode==1.3
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.55
ultralytics-thop==2.0.0
urllib3==2.2.2
uvicorn==0.30.6
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
